python: remove ability to build python2 modules for project
This commit removes Python 2 package generation ability from the project to encourage safer development with the Python 3 libraries instead.
Unfortunately documentation and tests require the Python 2 interpreter and may not work after these changes due to lack of Python 2 libs.
By default the project looks for the Python 3 libs and interpreter but can be overridden by the build config.

Signed-off-by: Mihai Tudor Panu <mihai.tudor.panu@intel.com>